Fix all api urls

+++ src/main/java/javatp/controller/AnswerController.java
@@ -1,6 +1,7 @@
 package javatp.controller;
 
 import java.util.List;
+
 import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,52 +15,77 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javatp.domain.Question;
 import javatp.domain.Answer;
 import javatp.service.AnswerService;
-
+import javatp.service.QuestionService;
 
 @RestController
-@RequestMapping(value = "/api/answers")
-public class AnswerController{
-    @Autowired
-    private AnswerService answerService;
-    
-	@PostMapping(value = "/answer")
-	public ResponseEntity<Object> createAnswer(@RequestBody Answer answer) {
+@RequestMapping(value = "/api/pois/{poiId}/questions/{questionId}/answers")
+public class AnswerController {
+	@Autowired
+	private AnswerService answerService;
+	@Autowired
+	private QuestionService questionService;
+
+	@PostMapping(value = "")
+	public ResponseEntity<Object> createAnswer(@PathVariable("poiId") long poiId,
+			@PathVariable("questionId") long questionId, @RequestBody Answer answer) {
+		if (!questionService.questionExistsByID(poiId, questionId))
+			throw new EntityNotFoundException();
+		answer.setQuestion(new Question(questionId));
 		Answer newAnswer = answerService.createAnswer(answer);
 		return ResponseEntity.ok(newAnswer);
-    }
-    
-    @GetMapping(value = "/answer/{id}")
-	public ResponseEntity<Object> getAnswer(@PathVariable("id") long id) {
-		Answer answer = answerService.getAnswer(id);
-		return ResponseEntity.ok(answer);
 	}
 
-    @PutMapping(value = "/answer/{id}")
-	public ResponseEntity<Object> updateAnswer(@PathVariable("id") long id, @RequestBody Answer answer) {
-		if(!answerService.AnswerExistsByID(id)) throw new EntityNotFoundException();
+	@GetMapping(value = "/{id}")
+	public ResponseEntity<Object> getAnswer(@PathVariable("poiId") long poiId,
+			@PathVariable("questionId") long questionId, @PathVariable("id") long id) {
+		if (!answerService.answerExistsByID(poiId, questionId, id))
+			throw new EntityNotFoundException();
+		return ResponseEntity.ok(answerService.getAnswer(id));
+	}
+
+	@PutMapping(value = "/{id}")
+	public ResponseEntity<Object> updateAnswer(@PathVariable("poiId") long poiId, @PathVariable("id") long id,
+			@PathVariable("questionId") long questionId, @RequestBody Answer answer) {
+		if (!answerService.answerExistsByID(poiId, questionId, id))
+			throw new EntityNotFoundException();
 		answer.setId(id);
+		answer.setQuestion(new Question(questionId));
 		Answer updatedAnswer = answerService.updateAnswer(answer);
 		return ResponseEntity.ok(updatedAnswer);
 	}
 
-	@DeleteMapping(value = "/answer/{id}")
-	public ResponseEntity<Object> deleteAnswer(@PathVariable("id") long id) {
-		Answer answer = answerService.getAnswer(id);
+	@DeleteMapping(value = "/{id}")
+	public ResponseEntity<Object> deleteAnswer(@PathVariable("poiId") long poiId,
+			@PathVariable("questionId") long questionId, @PathVariable("id") long id) {
+		if (!answerService.answerExistsByID(poiId, questionId, id))
+			throw new EntityNotFoundException();
+		Answer answer = new Answer(id);
 		answerService.deleteAnswer(answer);
 		return ResponseEntity.ok().build();
 	}
 
 	@GetMapping(value = "")
-	public ResponseEntity<Object> getAnswers() {
-        List<Answer> answers = answerService.getAllAnswers();
+	public ResponseEntity<Object> getAnswers(@PathVariable("poiId") long poiId,
+			@PathVariable("questionId") long questionId) {
+		if (!questionService.questionExistsByID(poiId, questionId))
+			throw new EntityNotFoundException();
+		List<Answer> answers = questionService.getQuestion(questionId).getAnswers();
 		return ResponseEntity.ok(answers);
 	}
 
-	@PostMapping(value = "")
-	public ResponseEntity<Object> createAnswers(@RequestBody List<Answer> answers) {
-        List<Answer> newAnswers = answerService.createAnswers(answers);
+	@PostMapping(value = "/batch")
+	public ResponseEntity<Object> createAnswers(@PathVariable("poiId") long poiId,
+			@PathVariable("questionId") long questionId, @RequestBody List<Answer> answers) {
+		if (!questionService.questionExistsByID(poiId, questionId))
+			throw new EntityNotFoundException();
+		Question question = new Question(questionId);
+		for (Answer answer : answers) {
+			answer.setQuestion(question);
+		}
+		List<Answer> newAnswers = answerService.createAnswers(answers);
 		return ResponseEntity.ok(newAnswers);
 	}
 }
